move client ContractClass to deprecated_contract_class
